+++ doc/change_log.md
@@ -1,5 +1,19 @@
 # Change Log
 
+**v0.5.0**
+- show home コマンドに下記オプションを追加
+    - screen name 引数: ログインユーザー以外のUserTimelineを取得可能にした
+    - --page オプション: limit数のツイートを page回分取得
+    - --since-id オプション: since-idより新しいツイートまで取得
+    - --max-id オプション: max-idと同じか古いツイートから取得
+    - --retry オプション: 通信エラー時、再チャレンジする回数。初期値 3 回
+- show friend, show follower コマンドを追加
+    - --limit, --page オプション: limit数のツイートを page回分取得
+    - --all オプション: friend, followerをすべて取得
+    - --as-id オプション: user idの一覧を取得
+    - --retry オプション: 通信エラー時、再チャレンジする回数。初期値 3 回
+- show home show tweet, show favorite コマンドの help 中 limit オプションに最大値を記載
+
 **v0.4.0**
 - show tweet コマンドに下記オプションを追加
     - screen name 引数: ログインユーザー以外のUserTimelineを取得可能にした